fastapi-backend/pyproject.toml
[tool.poetry]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = ["SrdjanTopic <64950418+SrdjanTopic@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "fastapi_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.9.4"
anyio = "3.6.2"
autopep8 = "2.0.1"
bcrypt = "4.0.1"
cffi = "1.15.1"
click = "8.1.3"
colorama = "0.4.6"
cryptography = "39.0.1"
ecdsa = "0.18.0"
fastapi = "0.92.0"
fastapi-sqlalchemy = "0.2.1"
greenlet = "2.0.2"
h11 = "0.14.0"
httptools = "0.5.0"
idna = "3.4"
mako = "1.2.4"
markupsafe = "2.1.2"
passlib = "1.7.4"
psycopg2 = "2.9.5"
psycopg2-binary = "2.9.5"
pyasn1 = "0.4.8"
pycodestyle = "2.10.0"
pycparser = "2.21"
pydantic = "1.10.5"
python-dotenv = "0.21.1"
python-jose = "3.3.0"
python-multipart = "0.0.5"
pyyaml = "6.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.4"
starlette = "0.25.0"
tomli = "2.0.1"
typing-extensions = "4.5.0"
uvicorn = "0.20.0"
watchfiles = "0.18.1"
websockets = "10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
